feat: add comprehensive analysis tabs for patient assessments
- Implemented AssessmentTab to display diagnostic evaluations, including clinical relationships between symptoms and diagnoses.
- Developed InsightsTab to present AI-generated risks, patterns, and suggestions for patient care.
- Created ObjectiveTab for showcasing objective evaluations and findings from physical exams and vital signs.
- Added PlanTab to outline treatment plans and recommendations, including follow-up suggestions.
- Introduced SubjectiveTab to summarize patient-reported symptoms and findings.
- Built SummaryTab to provide an overview of key diagnoses, symptoms, and clinical relationships, along with confidence scores and high-priority alerts.
- Defined types and interfaces in analysis-types.ts for enhanced data structure and type safety across components.
